requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
pydantic>=2.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0